+++ public/theme/boost_union/db/upgrade.php
@@ -921,6 +921,87 @@ function xmldb_theme_boost_union_upgrade($oldversion) {
         upgrade_plugin_savepoint(true, 2025100615, 'theme', 'boost_union');
     }
 
+    if ($oldversion < 2025100617) {
+        // Define fields email* to be added to theme_boost_union_menuitems.
+        $table = new xmldb_table('theme_boost_union_menuitems');
+        $field = new xmldb_field('email_cc', XMLDB_TYPE_TEXT, null, null, null, null, null, 'email');
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+        $field = new xmldb_field('email_bcc', XMLDB_TYPE_TEXT, null, null, null, null, null, 'email_cc');
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+        $field = new xmldb_field('email_subject', XMLDB_TYPE_TEXT, null, null, null, null, null, 'email_bcc');
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+        $field = new xmldb_field('email_body', XMLDB_TYPE_TEXT, null, null, null, null, null, 'email_subject');
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+
+        // Boost Union savepoint reached.
+        upgrade_plugin_savepoint(true, 2025100617, 'theme', 'boost_union');
+    }
+
+    if ($oldversion < 2025100618) {
+        // Define table theme_boost_union_flavours to be altered.
+        $table = new xmldb_table('theme_boost_union_flavours');
+
+        // Define field look_linkcolor to be added.
+        $field = new xmldb_field('look_linkcolor', XMLDB_TYPE_CHAR, '32', null, null, null, null, 'look_brandcolor');
+
+        // Conditionally launch add field look_linkcolor.
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+
+        // Define field look_buttonbrandcolor to be added.
+        $field = new xmldb_field(
+            'look_buttonbrandcolor',
+            XMLDB_TYPE_CHAR,
+            '32',
+            null,
+            null,
+            null,
+            null,
+            'look_linkcolor'
+        );
+
+        // Conditionally launch add field look_buttonbrandcolor.
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+
+        // Boost Union savepoint reached.
+        upgrade_plugin_savepoint(true, 2025100618, 'theme', 'boost_union');
+    }
+
+    if ($oldversion < 2025100619) {
+        // Define table theme_boost_union_flavours to be altered.
+        $table = new xmldb_table('theme_boost_union_flavours');
+
+        // Define field content_footnote to be added.
+        $field = new xmldb_field('content_footnote', XMLDB_TYPE_TEXT, null, null, null, null, null, 'look_navbarcolor');
+
+        // Conditionally launch add field content_footnote.
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+
+        // Define field content_footnote_format to be added.
+        $field = new xmldb_field('content_footnote_format', XMLDB_TYPE_INTEGER, '2', null, null, null, null, 'content_footnote');
+
+        // Conditionally launch add field content_footnote_format.
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+
+        // Boost Union savepoint reached.
+        upgrade_plugin_savepoint(true, 2025100619, 'theme', 'boost_union');
+    }
+
     // Load the builtin SCSS snippets into the database.
     // This is done with every plugin update, regardless of the plugin version.
     snippets::add_builtin_snippets();